Refactor codebase to conform to AGENTS.md
- Move `src/coreason_manifest/runtime/executor.py` to `src/coreason_manifest/utils/simulation_executor.py`.
- Move `src/coreason_manifest/interop/mcp.py` to `src/coreason_manifest/utils/mcp_adapter.py`.
- Delete `src/coreason_manifest/runtime/` and `src/coreason_manifest/interop/` directories.
- Update `src/coreason_manifest/__init__.py`, `src/coreason_manifest/cli.py` and tests to reflect the changes.
- Ensure strict adherence to the "Shared Kernel" architecture which forbids top-level runtime components.
