--- src/coreason_manifest/interop/__init__.py
@@ -1,13 +1,0 @@
-# Copyright (c) 2025 CoReason, Inc.
-#
-# This software is proprietary and dual-licensed.
-# Licensed under the Prosperity Public License 3.0 (the "License").
-# A copy of the license is available at https://prosperitylicense.com/versions/3.0.0
-# For details, see the LICENSE file.
-# Commercial use beyond a 30-day trial requires a separate license.
-#
-# Source Code: https://github.com/CoReason-AI/coreason-manifest
-
-from .mcp import CoreasonMCPServer, create_mcp_tool_definition
-
-__all__ = ["CoreasonMCPServer", "create_mcp_tool_definition"]
