+++ src/coreason_manifest/utils/mcp_adapter.py
